requirements.txt
psycopg2-binary==2.9.1
jinja2==3.1.2
cryptography==3.4.7